requirements.txt
# python                     3.11.13 
fastapi==0.116.1
pydantic==2.11.7
pydantic-core==2.33.2 
pymysql==1.1.1 
pymilvus==2.5.2
sqlalchemy==2.0.41
uvicorn==0.35.0
grpcio==1.67.1
grpcio-status==1.67.1
google-generativeai==0.8.5
httpx==0.28.1
weasyprint==61.2        # HTML to PDF 변환
jinja2==3.1.2          # 템플릿 엔진  
pillow==10.0.0         # 이미지 처리
